// src/apispec/src/models/CreateCMDContextRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sandbox0 API
 * Public HTTP APIs exposed by internal-gateway.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateCMDContextRequest
 */
export interface CreateCMDContextRequest {
    /**
     * 
     * @type {Array<string>}
     * @memberof CreateCMDContextRequest
     */
    command: Array<string>;
}

/**
 * Check if a given object implements the CreateCMDContextRequest interface.
 */
export function instanceOfCreateCMDContextRequest(value: object): value is CreateCMDContextRequest {
    if (!('command' in value) || value['command'] === undefined) return false;
    return true;
}

export function CreateCMDContextRequestFromJSON(json: any): CreateCMDContextRequest {
    return CreateCMDContextRequestFromJSONTyped(json, false);
}

export function CreateCMDContextRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateCMDContextRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'command': json['command'],
    };
}

export function CreateCMDContextRequestToJSON(json: any): CreateCMDContextRequest {
    return CreateCMDContextRequestToJSONTyped(json, false);
}

export function CreateCMDContextRequestToJSONTyped(value?: CreateCMDContextRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'command': value['command'],
    };
}