// src/apispec/src/models/PTYSize.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sandbox0 API
 * Public HTTP APIs exposed by internal-gateway.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PTYSize
 */
export interface PTYSize {
    /**
     * 
     * @type {number}
     * @memberof PTYSize
     */
    rows?: number;
    /**
     * 
     * @type {number}
     * @memberof PTYSize
     */
    cols?: number;
}

/**
 * Check if a given object implements the PTYSize interface.
 */
export function instanceOfPTYSize(value: object): value is PTYSize {
    return true;
}

export function PTYSizeFromJSON(json: any): PTYSize {
    return PTYSizeFromJSONTyped(json, false);
}

export function PTYSizeFromJSONTyped(json: any, ignoreDiscriminator: boolean): PTYSize {
    if (json == null) {
        return json;
    }
    return {
        
        'rows': json['rows'] == null ? undefined : json['rows'],
        'cols': json['cols'] == null ? undefined : json['cols'],
    };
}

export function PTYSizeToJSON(json: any): PTYSize {
    return PTYSizeToJSONTyped(json, false);
}

export function PTYSizeToJSONTyped(value?: PTYSize | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'rows': value['rows'],
        'cols': value['cols'],
    };
}