// src/apispec/src/models/ProcessType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sandbox0 API
 * Public HTTP APIs exposed by internal-gateway.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ProcessType = {
    Repl: 'repl',
    Cmd: 'cmd'
} as const;
export type ProcessType = typeof ProcessType[keyof typeof ProcessType];


export function instanceOfProcessType(value: any): boolean {
    for (const key in ProcessType) {
        if (Object.prototype.hasOwnProperty.call(ProcessType, key)) {
            if (ProcessType[key as keyof typeof ProcessType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ProcessTypeFromJSON(json: any): ProcessType {
    return ProcessTypeFromJSONTyped(json, false);
}

export function ProcessTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProcessType {
    return json as ProcessType;
}

export function ProcessTypeToJSON(value?: ProcessType | null): any {
    return value as any;
}

export function ProcessTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): ProcessType {
    return value as ProcessType;
}